requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.0.0
azure-identity==1.19.0
httpx<0.28.0
pydantic==2.5.0
